session6.py
import time

def checker(fn):
    """
    A closure that checks if a function has a docstring of sufficient length before executing it.

    This function verifies that the passed function `fn` has a docstring that is more than 50 characters
    long (excluding spaces). If the docstring meets this requirement, the function is executed. Otherwise,
    a ValueError is raised.

    Args:
        fn (function): The function to be checked and potentially executed.

    Raises:
        TypeError: If the passed argument is not a function or is a class type.
        ValueError: If the function's docstring is missing or shorter than 50 characters.

    Returns:
        function: The inner function that enforces the docstring length check before calling the original function.

    Example:
        @checker
        def example_function():
            \"\"\"This is an example function with a sufficiently long docstring.\"\"\"
            return "Function executed"

        print(example_function())  # Output: The function example_function has a description of more than 50 characters in its docstring.
    """
    if not callable(fn):
        raise TypeError("The passed argument is not a function")
    if isinstance(fn, type):
        raise TypeError("The passed argument is a class type. Please use this closure only for functions")

    chars = 50

    def enforce_docstring_length(*args, **kwargs):
        """
        Checks the length of the docstring and executes the function if the docstring is sufficiently long.

        This function checks if the docstring of the function `fn` is longer than 50 characters (excluding spaces).
        If the docstring is sufficiently long, the function is executed with the provided arguments. If not, a ValueError is raised.

        Raises:
            ValueError: If the function's docstring is missing or shorter than 50 characters.

        Returns:
            The return value of the original function `fn` if it passes the docstring length check.
        """
        doc_string = fn.__doc__
        if doc_string:
            doc_string = doc_string.replace(" ", "")
            if len(doc_string) < chars:
                raise ValueError(f"Function '{fn.__name__}' requires a docstring longer than 50 characters.")
            else:
                print(f"The function {fn.__name__} has a description of more than {chars} characters in its docstring.")
                return fn(*args, **kwargs)
        else:
            raise ValueError(f"The passed function has no docstring.")

    return enforce_docstring_length

def fibonacci_closure():
    """
    Returns a closure that generates the next Fibonacci number each time it is called.

    This closure initializes with two initial values, `a` and `b`. Each time the
    `next_fibonacci_number` function is called, it calculates the next Fibonacci
    number based on the previous two numbers stored in `a` and `b`. The sequence
    starts from 0.

    Returns:
        function: A closure that returns the next Fibonacci number when called.

    Example:
        fib = fibonacci_closure()
        print(fib())  # Output: 0
        print(fib())  # Output: 1
        print(fib())  # Output: 1
        print(fib())  # Output: 2
        print(fib())  # Output: 3
    """
    a, b = 0, 1

    def next_fibonacci_number(*args, **kwargs):
        """
        Calculates the next Fibonacci number based on the previous two values.

        This function does not accept any arguments. It raises a ValueError if
        any arguments are passed.

        Raises:
            ValueError: If any arguments are passed to the function.

        Returns:
            int: The next Fibonacci number in the sequence.
        """
        if args or kwargs:
            raise ValueError("No arguments should be passed to this function.")

        nonlocal a, b
        next_val = a
        a, b = b, a + b
        return next_val

    return next_fibonacci_number

func_count={}

def function_counter_with_one_dict(fn):
    """
    A decorator that counts how many times a function is called and updates a global dictionary with the counts.

    Args:
        fn (function): The function to be decorated and counted.

    Raises:
        TypeError: If the passed argument is not a function or if it is a class type.

    Returns:
        function: The inner function that wraps the original function and keeps track of the call count.

    Example:
        @counter_with_one_dict
        def add(a, b):
            return a + b

        print(add(1, 2))  # Output: add has been called 1 times
        print(add(3, 4))  # Output: add has been called 2 times
        print(func_count)  # Output: {'add': 2}
    """
    if not callable(fn):
        raise TypeError("The passed argument is not a function")
    if isinstance(fn, type):
        raise TypeError("The passed argument is a class type. Please use this closure only for functions")

    cnt = 0  # initially fn has been run zero times

    def inner(*args, **kwargs):
        nonlocal cnt
        global func_count
        inner.__name__ = fn.__name__
        inner.__doc__ = fn.__doc__
        cnt += 1
        func_count[fn.__name__] = cnt
        print('{0} has been called {1} times'.format(fn.__name__, cnt))
        return fn(*args, **kwargs)

    def reset_counter():
        nonlocal cnt
        cnt = 0
    inner.reset_counter = reset_counter

    return inner

def function_counter_multi_dict(counter_dict):
    """
    A decorator factory that counts how many times a function is called and updates a specified dictionary with the counts.

    Args:
        counter_dict (dict): The dictionary to be updated with the function call counts.

    Raises:
        TypeError: If the passed argument is not a function or if it is a class type.

    Returns:
        function: A decorator that wraps the original function and keeps track of the call count.

    Example:
        func_count = {}
        @function_counter_multi_dict(func_count)
        def add(a, b):
            return a + b

        print(add(1, 2))  # Output: add has been called 1 times
        print(add(3, 4))  # Output: add has been called 2 times
        print(func_count)  # Output: {'add': 2}
    """
    if not isinstance(counter_dict, dict):
        raise TypeError("The counter_dict argument must be a dictionary")

    def decorator(fn):
        if not callable(fn):
            raise TypeError("The passed argument is not a function")
        if isinstance(fn, type):
            raise TypeError("The passed argument is a class type. Please use this decorator only for functions")

        cnt = 0  # initially fn has been run zero times

        def inner(*args, **kwargs):
            nonlocal cnt
            inner.__name__ = fn.__name__
            inner.__doc__ = fn.__doc__
            cnt += 1
            counter_dict[fn.__name__] = cnt
            print('{0} has been called {1} times'.format(fn.__name__, cnt))
            return fn(*args, **kwargs)

        def reset_counter():
            """
            This resets the counter on the functions to zero.
            """
            nonlocal cnt
            cnt = 0
        inner.reset_counter = reset_counter

        return inner

    return decorator
